Implement new key calculations
Adds cascading key calculations where it can recursively find the
challenge key by doing bfs across a list of linked lists basically.
It is less than ideal, and should really be a graph solution.

However, this fixes a major hurdle with the startingMode, and a bunch of
strings being passed around.